Rename the generic fallback JSON to include the `cp` prefix
When an image that is past a checkpoint request the generic fallback
JSON, we need to include which checkpoint number that image is past.
Otherwise, we would not be able to confidently propose a valid update.

This information was encoded as `.<checkpoint_number>.json`. But that
value could have been easily mistaken with the image buildid.

To be more explicit that this number is indicating a checkpoint, we
prefix `cp` to it. E.g. the generic fallback meta JSON for images that
are past the checkpoint 2 and that require the steamdeck beta variant
will be `steamdeck-beta.cp2.json`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>