tests: Remove images with the same version, release and buildid
The server is using the version, release and buildid information to
pinpoint what is the actual image a client is currently using.
However, the server might not reliably distinguish the images if they
share the same values.

With this commit we remove duplicated images from the automated tests
and, as a follow up, we will put some check in-place to actually forbid
this situation to ever happen.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>