Preparatory changes to support the new "branch" concept
With this commit we add the initial essential pieces to separate the
concept of variants and update channels.

Moving forward, the variant will just encode the "flavor" of the image,
e.g. "steamdeck". Instead, the update channel, will be encoded into a
separated "branch" variable.

For example new Steam Deck images will have a manifest similar to this
one:
{
  "product": "steamos",
  "release": "holo",
  "variant": "steamdeck",
  "branch": "beta",
  "arch": "amd64",
  "version": "3.6.1",
  "buildid": "20240112.100"
}

With this commit we separate the variant into variant+branch and add
backward compatibility to ensure that older images in the server pool
continue to be supported.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>