Store new image JSON manifests into a separate path
With the introduction of branches we need to encode this information
into the update meta path.

Given the need to change the path, we can take this opportunity to also
perform a refactoring.
One problem with the old path was that the "release" was not included at
all.
Another minor issue was that the generic fallback was just one level up,
meaning that we ended up with one generic JSON for each "version", even
if they were essentially all the same. E.g. "[...]/3.5.0/steamdeck.json"
and "[...]/3.5.1/steamdeck.json" and so on.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>