teststaticserver: Do not trigger a refresh when not needed
When we change the expectations, we want to test that the server doesn't
automatically re-trigger a rescan until we actually touch the
`updated.txt` file.
Before this commit we were unexpectedly touching the `updated.txt` file.

The tests usually succeeded out of timing luck. However, from time to
time, the automated tests failed because the meta output data changed
before `diff` was able to run.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>