testclient: Add coverage for the meta and query download
Add automated tests to ensure `download_update_from_*_url()` methods are
working as expected.

Instead of deploying a local temporary web server, we reuse the
instances that we have in production. This allows us to have tests that
are as close to the reality as possible.
It is safe to use the production endpoints because they are expected to
always be there in the future, otherwise that would break the update
system for the Steam Decks that are already in the wild.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>