client: Remove QueryUrl support
The old dynamic server is planned to be stopped with jupiter/tasks#448
and the queries will be redirected to the generic response
`steamdeck.json`.

Given that Jupiter already made the switch to the meta URL, there is no
reason to keep the code to support the deprecated query URL around.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>