image: Use the fallback when switching variant
When we request to switch to a completely different variant, we should
go directly to the generic fallback meta JSON.

We don't expect the new variant to have the current image
`<version>/<buildid>.json` anyway.

Furthermore, if `<version>/<buildid>.json` was actually available, that
would mean that the requested variant had a image with the same
"version" and "build id". So that JSON was not supposed to be for us in
the first place.
This is because we don't enforce unique version+buildid across all
variants.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>