client: Remove old TODOs
The client side should be as simple as possible. If we make it
"smarter", it will inevitably increase the complexity of supporting that
system down the road.
Clients should be supported "forever", so we should only stick to the
essential operations and leave as much logic as possible to the server
side.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>