Remove legacy dynamic server
The new static server was introduced to eventually replace the legacy
dynamic server.

The dynamic server is not covered by our automated tests and this poses
a continuous support burden when applying changes to steamos-atomupd.
It is always tricky to ensure the correctness of the dynamic server and
there is a concrete risk of having two server instances that diverge.

Jupiter made the switch to the static server since the version 3.5,
which is currently in the stable branch.

Backward compatibility with the Steam Decks that are still using older
images, is ensured by redirecting requests to the static content
provided by the static server.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>