Add StrictPoolValidation config field
When bootstrapping a new server it could very likely happen to not have
an image for every variants and branches listed in the server
configuration.
In those cases, technically it is not an error but just the expected
behavior until the server becomes fully operational.

To cover this use case we add a new `StrictPoolValidation`, to
explicitly tell the server that it should relax its checks because it is
Okay not have images for some of the branches or variants.

In the future we might expand the number of strict checks that are gated
behind `StrictPoolValidation`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>