Revamp the "release" and minor/major update concepts
The concept of "releases" being an alphabetically ordered list to
differentiate major upgrades is not necessary anymore. With the
introduction of numbered checkpoints, we now have a proper system to
handle OS updates that introduce breaking changes.

For this reason, with this commit we deprecate the distinction between
"minor" and "major" upgrades. And the "release" field of a build is now
used to only differentiate different OS "families".

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>