client: Open the JSON URL directly without using the filesystem
When asking the server for a JSON, we were storing the response to the
filesystem for historic reasons.

This could have caused problems if we executed multiple
`steamos-atomupd-client --query-only` concurrent instances, because
they'd all try to manipulate the same `update.json` file.

Instead of using a mutex, we can simply read the URL response directly
into a variable and skip the whole writing to the filesystem part.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>