client: Continue even if we can't estimate the download size
If another process is already using rauc, we'll not be able to create an
index file for the active rootfs, because rauc can't be executed
concurrently.

This situation can be reproduced by launching `steamos-atomupd-client`
in one terminal and, while it is still applying an update, launching
`steamos-atomupd-client --query-only --estimate-download-size` in
another terminal.
The second command will fail with a `CalledProcessError` exception.

Given that estimating the download size is not a critical operation, we
should not abort the entire query procedure if it happens that we can't
perform the estimation. This is also in line with what we already do in
`estimate_download_size()` regarding the other possible errors that
might appear.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>