mkmanifest: Drop the `estimate_size` key from image manifests
The `estimated_size` key is only relevant in the JSON update candidates.
In a manifest that describes a single image, that key is always expected
to be zero because the update size depends on which base image a client
already has installed.

With this commit we drop that key to avoid any possible confusion.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>