Introduce the "default_update_branch" manifest field
So far the `branch` field inside an image manifest was used as its
default fallback choice, until the user manually selected which channel
to follow.

However, we now have a use case where developers want to create images
that defaults to a branch that is different from the one their are
currently in. For example a beta image that defaults to the stable
channel.

In that use case, using the `branch` field to convey the "default
branch" info can be misleading.
For this reason, this commit introduces a new optional field called
`default_update_branch`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>