Add ability to mark a variant as End-of-life
When developers use custom variants for their short lived smoke tests,
having a system that can automatically redirect EOL variants to their
replacement can be handy.

With the new option VariantsEOL, we can list which variants terminated
their purpose and what's the replacement that should be installed
instead.

The new optional entry `replacement_eol_variant` is used by the server
to inform the clients that the variant that was initially requested is
now EOL. And the value of that entry is what should be tracked instead.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>