client: Show a warning that this is probably not the right CLI to use
When using `steamos-atomupd-client` without being explicit about which
update we are interested in, it might end up proposing an unexpected
update.

This happens because `steamos-atomupd-client` is a low level CLI that
doesn't parse the `preferences.conf` of the atomupd D-Bus API.
This means that, for example, if a user is in a `stable` OS and chooses
`beta` from the Steam client, launching `steamos-atomupd-client`,
without additional arguments, will still look for `stable` updates.

With this commit we print a warning to inform developers that in the
vast majority of cases the CLI they should be using is
`atomupd-manager`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>