Allow a more granular branch order
While a linear branches order works fine for most use cases, it doesn't
cover the special-purpose branches, such as `bc` and `rc`.
In those cases, the images are used only for a quick smoketesting before
being promoted. This means that we don't want to propose, for example, a
new `rc` build to users that are following the `beta` branch.

With this commit we add support for a more granular branch order, where
it is possible to independently specify which branches to consider for
each one of them.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>