Remove the option to set a custom manifest in client.conf
We remove this option because there aren't any real use cases where it
is desired to point `steamos-atomupd` to a custom image manifests, apart
from developers when doing some manual testings.
However, in those cases it is still possible to simply edit
`/etc/steamos-atomupd/manifest.json` or launch `steamos-atomupd-client`
with the `--manifest-file` option.

This client.conf field was never used in production, so it safe to be
removed.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>