Introduce option to generate the new "remote-info.conf" file
With the latest changes to the atomupd-daemon, clients are now able to
ask the server for a more up-to-date list of known branches and
variants.

This can be useful to update the hardcoded list that clients have, e.g.
when we recently introduced the "preview" and "pc" branches.

The new option `GenerateRemoteInfoConfig` can be used to let the server
automatically generate those files.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>