Generate the remote info config from a user provided list
There are cases where it's not desirable to automatically generate the
`remote-info.conf` files from the main variants and branches lists.

For example in Jupiter we currently have the `staging` branch, which is
a special purpose branch used primarily for quickly testing unstable
changes before promoting them to main. That branch should be hidden from
the UI and only selectable using the CLI.

Another special case is for the EOL variants. Users running an EOL
variant should still see their own variant in the list of choices, while
for all the other images there is no point in showing them.

To support those use cases we add a new optional section
`Images.ProvideRemoteInfoConfig`, where we can explicitly list the
variants and branches that end users should know about.

Eventual EOL variants listed there, will be automatically handled as
expected (shown for the EOL images but hidden for all the others).

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>