staticserver: Estimate the download size only when really necessary
Estimating the download size is a quite fast operation, but it requires
calling an external process (rauc) and the context switch time adds up
quickly when going through a large number of images.

Now we only estimate the download size if the update JSON actually
changed from what we already had in the filesystem.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>