Convert "ProvideRemoteInfoConfig" to an arch specific property
The static server was designed from the beginning to, in theory, support
multiple architectures with a single instance. While this feature was
never used, we should still try to not regress it.

The new `ProvideRemoteInfoConfig` option was affecting all
architectures, regardless of the declared arches. To make it more
flexible, now that option requires a `.$ARCH` suffix, to allow setting
different values of variants and branches for each architecture.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>