client: Don't generate the index, if missing, while checking for updates
Under normal circumstances we always expect the index file to be already
available. The only two edge cases where this is not happening, is after
a factory reset and for the very fist update in OOBE.

Generating from scratch the index file for the rootfs can take up to 30
seconds. When checking for available updates we should not stall the
process for that long. Instead we should keep the generation only when
actually applying an update.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>