utils: Set a reasonable timeout when extracting a raucb
Extracting a local raucb should be nearly instantaneous, while
extracting a remote raucb should take only up to a few seconds (time
needed to download a 5MB file).

We set a timeout of 20 seconds after which we stop the extract process
and continue. This ensures that we don't needlessly slowdown the "check
for updates" operation.

Especially considering that, under normal circumstances, the server is
already providing the estimated download size, and this is only a
fallback operation.

If the server didn't provide the pre-calculated estimation, and we hit
the 20 seconds timeout, the "check for updates" result will be left
without the info about the update size.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>